Set AndroidJunit4 test runner
