+++ sample/build.gradle
@@ -8,6 +8,8 @@ android {
         applicationId "com.example.snapper"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
+
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
@@ -15,6 +17,10 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
         }
+    }
+
+    packagingOptions {
+        exclude 'LICENSE.txt'
     }
 }
 
@@ -32,9 +38,5 @@ dependencies {
     compile 'com.jakewharton:butterknife:6.1.0'
     compile 'com.jakewharton.timber:timber:2.7.1'
     //
-    testCompile 'junit:junit:4.12'
-    testCompile 'org.mockito:mockito-core:2.+'
-    testCompile 'org.robolectric:robolectric:3.+'
-    testCompile 'org.robolectric:shadows-support-v4:3+'
-    testCompile 'org.apache.maven:maven-ant-tasks:2.1.3'
+    androidTestCompile 'com.android.support.test:testing-support-lib:0.1'
 }
